OpenTabList getTabsForDatabase now returns a list rather than an array, which can be empty, but never null.
Changed the CursorManager to accept a name - the name of the thread/object that is setting / clearing the hourglass cursor. Improved the stuck hourglass detector, as doing thread dumps wasn't particularly revealing - also remove entries from the list of hourglass setters when the cursot returns to normal.

Changed the SnailActionListener so that it is now responsible for setting the cursor back to normal after the lengthy operation.